fix pico-dev

+++ bsp/pico-dev/pikascript/pikascript-api/PikaStdLib_MemChecker.h
@@ -9,8 +9,8 @@
 
 PikaObj *New_PikaStdLib_MemChecker(Args *args);
 
-float PikaStdLib_MemChecker_getMax(PikaObj *self);
-float PikaStdLib_MemChecker_getNow(PikaObj *self);
+double PikaStdLib_MemChecker_getMax(PikaObj *self);
+double PikaStdLib_MemChecker_getNow(PikaObj *self);
 void PikaStdLib_MemChecker_max(PikaObj *self);
 void PikaStdLib_MemChecker_now(PikaObj *self);
 void PikaStdLib_MemChecker_resetMax(PikaObj *self);
